working old h5md tests
